requirements.txt
boto3==1.4.7
botocore==1.7.9
docutils==0.14
jmespath==0.9.3
python-dateutil==2.6.1
PyYAML==3.12
s3transfer==0.1.11
six==1.10.0
